fix: reviewer comments
